% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_age.R
\name{covid_age}
\alias{covid_age}
\title{Extract COVID-19 age data from source}
\usage{
covid_age()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from OpenDataPhilly.org which provides .csv files taken from a CARTO SQL database API.

\code{covid_age} pulls positive COVID-19 aggregate age breakdown data from the COVID Tests and Cases registry as well as the COVID Deaths registry and presents the output in a pre-processed, tidy dataset.

Note that this extraction provides only aggregate summation of data from the start of recording to the ETL \code{TimeStamp}.
}
\references{
COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}
COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}
}
