% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_date.R
\name{covid_date}
\alias{covid_date}
\title{Extract COVID-19 date data from source}
\usage{
covid_date()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from data from \url{https://www.opendataphilly.org} which provides .csv files taken from a CARTO SQL database API.

\code{covid_date} pulls test result and date data from the COVID Tests and Cases registry as well as the COVID Deaths registry and presents the output in a pre-processed, tidy dataset.
}
\references{
- COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}

- COVID Tests by Date: \url{https://www.opendataphilly.org/dataset/covid-cases/resource/85fead1a-0ec3-4856-98a9-27ba4fa941ee}

- COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}

- COVID Deaths by Date: \url{https://www.opendataphilly.org/dataset/covid-deaths/resource/45760e22-e2de-456f-ad0f-6c491b0f2e42}
}
