% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_gender.R
\name{covid_gender}
\alias{covid_gender}
\title{Extract COVID-19 gender breakdown data from source}
\usage{
covid_gender()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from data from \url{https://www.opendataphilly.org} which provides .csv files taken from a CARTO SQL database API.

\code{covid_gender} pulls data from the COVID Tests and Cases registry and presents the output in a pre-processed, tidy dataset.

Note that this extraction provides only aggregate summation of data from the start of recording to the ETL \code{TimeStamp} which generally takes place at 15:00:00 UTC each day.
}
\references{
- COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}

- COVID Cases by Sex: \url{https://www.opendataphilly.org/dataset/covid-cases/resource/856a0639-334c-43c2-a53b-1b3f5611fbda}

- COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}

- COVID Deaths by Age: \url{https://www.opendataphilly.org/dataset/covid-deaths/resource/33e68765-d07a-410a-9736-ac28fd587ce6}
}
