% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_hospitalizations.R
\name{covid_hospitalizations}
\alias{covid_hospitalizations}
\title{Extract COVID-19 hospitalization data from source}
\usage{
covid_hospitalizations()
}
\description{
\code{covid_hospitalizations()} returns a data set in list format pertaining to the various COVID Hospitalizations exports on OpenDataPhilly in list format. This package is not responsible for the raw dataset or ETL process from OpenDataPhilly. ETL for OpenDataPhilly is done on a daily basis.

Accessible sub-datasets include:
\itemize{
\item{age}
\item{date}
\item{race}
\item{sex}
\item{zip}
}

\strong{Note}: For aggregate data, counts less than 6 patients are rounded up to 6 to preserve patient health information anonyimity.
}
\references{
\href{https://www.opendataphilly.org/dataset/covid-hospitalizations}{COVID Hospitalizations}

\href{https://www.opendataphilly.org/dataset/covid-hospitalizations/resource/fc4d4c7d-392b-4da7-8a9e-cec0ccea8b6f}{COVID Hospitalizations by Age}

\href{https://www.opendataphilly.org/dataset/covid-hospitalizations/resource/be0f3aa9-c4d5-461d-ac55-245872de69ba}{COVID Hospitalizations by Date}

\href{https://www.opendataphilly.org/dataset/covid-hospitalizations/resource/9335e9ea-e1ca-46ce-bbcd-ac1d708cb125}{COVID Hospitalizations by Race}

\href{https://www.opendataphilly.org/dataset/covid-hospitalizations/resource/78b9a4e4-daef-464b-a883-91efccef86f7}{COVID Hospitalizations by Sex}

\href{https://www.opendataphilly.org/dataset/covid-hospitalizations/resource/a73874d4-311c-489f-a976-9bc0bf2942d2}{COVID Hospitalizations by ZIP}
}
