% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_outcome.R
\name{covid_outcome}
\alias{covid_outcome}
\title{Extract COVID-19 outcomes data from source}
\usage{
covid_outcome()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from data from \url{https://www.opendataphilly.org} which provides .csv files taken from a CARTO SQL database API.

\code{covid_outcome} pulls data from the COVID Tests and Cases registry and presents the output in a pre-processed, tidy dataset.

Note that this extraction provides only aggregate summation of data from the start of recording to the ETL \code{TimeStamp} which generally takes place at 15:00:00 UTC each day.
}
\references{
- COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}

- COVID Tests by Outcome: \url{https://www.opendataphilly.org/dataset/covid-cases/resource/cdf4a96d-3655-49bc-8490-2a73260a29a7}
}
