% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_race.R
\name{covid_race}
\alias{covid_race}
\title{Extract COVID-19 race breakdown data from source}
\usage{
covid_race()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from data from \url{https://www.opendataphilly.org} which provides .csv files taken from a CARTO SQL database API.

\code{covid_race} pulls data from the COVID Tests and Cases registry and presents the output in a pre-processed, tidy dataset. This data does not include race breakdown for patients who died. That is handled using \code{covid_race_died}.

Note that this extraction provides only aggregate summation of data from the start of recording to the ETL \code{TimeStamp} which generally takes place at 15:00:00 UTC each day.
}
\references{
- COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}

- COVID Cases by Race: \url{https://www.opendataphilly.org/dataset/covid-cases/resource/a83adfe0-4dac-4718-9996-dd06a9a67192}

- COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}

- COVID Death by Race" \url{https://www.opendataphilly.org/dataset/covid-deaths/resource/c053a703-d99c-4003-93ca-caa5db17369a}
}
