% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_race.R
\name{covid_race_died}
\alias{covid_race_died}
\title{Extract COVID-19 race breakdown data from source}
\usage{
covid_race_died()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from data from \url{https://www.opendataphilly.org} which provides .csv files taken from a CARTO SQL database API.

\code{covid_race_died} pulls data from the COVID Tests and Cases registry and presents the output in a pre-processed, tidy dataset. This data only includes patients who died, for patients who survived use \code{covid_race}.

Note that this extraction provides only aggregate summation of data from the start of recording to the ETL \code{TimeStamp} which generally takes place at 15:00:00 UTC each day.
}
\references{
- COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}
- COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}
}
