% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_tests_cases.R
\name{covid_tests_cases}
\alias{covid_tests_cases}
\title{Extract COVID-19 tests and cases data from source}
\usage{
covid_tests_cases()
}
\description{
\code{covid_tests_cases()} returns a data set in list format pertaining to the various COVID Tests and Cases as well as the COVID Deaths exports on OpenDataPhilly in list format. This package is not responsible for the raw dataset or ETL process from OpenDataPhilly. ETL for OpenDataPhilly is done on a daily basis.

Accessible sub-datasets include:
\itemize{
\item{age}
\item{date}
\item{outcome}
\item{race}
\item{sex}
\item{zip}
}

\strong{Note}: For aggregate data, counts less than 6 patients are rounded up to 6 to preserve patient health information anonyimity.

\strong{Note}: The author of this package determined it would be best to combine elements of the Deaths and Tests and Cases datasets for ease of use here.
}
\references{
\href{https://www.opendataphilly.org/dataset/covid-hospitalizations}{COVID Tests and Cases}

\href{https://www.opendataphilly.org/dataset/covid-deaths}{COVID Deaths}

\href{https://www.opendataphilly.org/dataset/covid-cases/resource/86d7c7f5-b348-4ce2-bbb3-8d8e6f6dec98}{COVID Tests and Cases by Age}

\href{https://www.opendataphilly.org/dataset/covid-deaths/resource/33e68765-d07a-410a-9736-ac28fd587ce6}{COVID Deaths by Age}

\href{https://www.opendataphilly.org/dataset/covid-cases/resource/85fead1a-0ec3-4856-98a9-27ba4fa941ee}{COVID Tests and Cases by Date}

\href{https://www.opendataphilly.org/dataset/covid-deaths/resource/45760e22-e2de-456f-ad0f-6c491b0f2e42}{COVID Deaths by Date}

\href{https://www.opendataphilly.org/dataset/covid-hospitalizations/resource/9335e9ea-e1ca-46ce-bbcd-ac1d708cb125}{COVID Tests and Cases by Race}

\href{https://www.opendataphilly.org/dataset/covid-cases/resource/cdf4a96d-3655-49bc-8490-2a73260a29a7}{COVID Tests and Cases by Outcome}

\href{https://www.opendataphilly.org/dataset/covid-deaths/resource/c053a703-d99c-4003-93ca-caa5db17369a}{COVID Deaths by Race}

\href{https://www.opendataphilly.org/dataset/covid-cases/resource/856a0639-334c-43c2-a53b-1b3f5611fbda}{COVID Tests and Cases by Sex}

\href{https://www.opendataphilly.org/dataset/covid-cases/resource/d4d1e48a-d401-405c-972b-c45292c3d4f5}{COVID Tests and Cases by ZIP}

\href{https://www.opendataphilly.org/dataset/covid-deaths/resource/7da31c98-c2ff-4fea-b313-2a40230d4810}{COVID Deaths by ZIP}
}
