% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_zip.R
\name{covid_zip}
\alias{covid_zip}
\title{Extract COVID-19 date data from source}
\usage{
covid_zip()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from data from \url{https://www.opendataphilly.org} which provides .csv files taken from a CARTO SQL database API.

\code{covid_zip} pulls test result and death data per date and ZIP code from the COVID Tests and Cases registry as well as the COVID Deaths registry and presents the output in a pre-processed, tidy dataset.
}
\references{
- COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}

- COVID Tests by ZIP: \url{https://www.opendataphilly.org/dataset/covid-cases/resource/d4d1e48a-d401-405c-972b-c45292c3d4f5}

- COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}

- COVID Deaths by ZIP: \url{https://www.opendataphilly.org/dataset/covid-deaths/resource/7da31c98-c2ff-4fea-b313-2a40230d4810}
}
