% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_gender.R
\name{covid_gender}
\alias{covid_gender}
\title{Extract COVID-19 gender breakdown data from source}
\usage{
covid_gender()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from OpenDataPhilly.org which provides .csv files taken from a CARTO SQL database API.

\code{covid_gender} pulls data from the COVID Tests and Cases registry and presents the output in a pre-processed, tidy dataset.

Note that this extraction provides only a sum total since the start of recording and not a summary over time.
}
\references{
COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}
COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}
}
