% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_race.R
\name{covid_race}
\alias{covid_race}
\title{Extract COVID-19 race breakdown data from source}
\usage{
covid_race()
}
\description{
The \code{covid_*} family of functions pull COVID-19 tracking
data from OpenDataPhilly.org which provides .csv files taken from a CARTO SQL database API.

\code{covid_race} pulls data from the COVID Tests and Cases registry and presents the output in a pre-processed, tidy dataset. This data does not include race breakdown for patients who died. That is handled using \code{covid_race_died}.

Note that this extraction provides only a sum total since the start of recording and not a summary over time.
}
\references{
- COVID Tests and Cases: \url{https://www.opendataphilly.org/dataset/covid-cases}
- COVID Deaths: \url{https://www.opendataphilly.org/dataset/covid-deaths}
}
